refactor: #1 core-navigator 설계 - base_mvvm_bottom_sheet_bridge BottomSheet Migrate 작업
