+++ features/base-mvvm/src/main/res/layout/a_refactor_base_test.xml
@@ -46,6 +46,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="10dp"
             android:background="#8E3939"
+            android:onClick="@{()->vm.moveToBottomSheet()}"
             android:padding="15dp"
             android:text="일반 바텀 시트"
             android:textColor="#FFF" />
@@ -56,6 +57,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="10dp"
             android:background="#C8964B"
+            android:onClick="@{()->vm.moveToShareBottomSheet()}"
             android:padding="15dp"
             android:text="공유 바텀 시트"
             android:textColor="#FFF" />
