+++ features/base_mvvm_bottom_sheet/src/main/java/com/features/base_mvvm_bottom_sheet/di/FeatureModule.kt
@@ -1,8 +1,6 @@
 package com.features.base_mvvm_bottom_sheet.di
 
-import com.features.base_mvvm_bottom_sheet.impl.BaseMvvmBottomSheetBridgeImpl
 import com.features.base_mvvm_bottom_sheet.source.ApiService
-import com.features.base_mvvm_bottom_sheet_bridge.BaseMvvmBottomSheetBridge
 import com.hmju.core.network.NetworkProvider
 import dagger.Module
 import dagger.Provides
@@ -17,15 +15,11 @@ import dagger.hilt.components.SingletonComponent
 @InstallIn(SingletonComponent::class)
 @Module
 internal object FeatureModule {
-    @Provides
-    fun provideBridge(): BaseMvvmBottomSheetBridge {
-        return BaseMvvmBottomSheetBridgeImpl()
-    }
 
     @Provides
     fun provideApiService(
-        provider: NetworkProvider
+	provider: NetworkProvider
     ): ApiService {
-        return provider.createApiService(ApiService::class.java)
+	return provider.createApiService(ApiService::class.java)
     }
 }
